teams and invitations should be working properly

+++ app/controllers/invitations_controller.php
@@ -28,23 +28,62 @@ class InvitationsController extends AppController {
 			$this->redirect(array('action'=>'index'));
 			
 		}
-		$this->Invitation->Team->bindModel(array('hasMany'=>array('UsersTeams')));
-		$this->Invitation->Team->UsersTeams->create();
-		$this->data['UsersTeams']['team_id'] =  $invitation['Invitation']['team_id'];
-		$this->data['UsersTeams']['user_id'] = $invitation['Invitation']['user_id'];
-		$this->Invitation->delete();
-		if($this->Invitation->Team->UsersTeams->save ($this->data))
-		{
-			$this->Session->setFlash('Invitation accepted!');
-			$this->redirect(array('controller'=>'teams','action'=>'view', $this->data['UsersTeams']['team_id']));
+		$members = $this->Invitation->Team->find('all', array(
+							'conditions'=>array('Team.id'=>$invitation['Invitation']['team_id']),
+							'contain'=>array(
+								
+								'User'=> array('fields' => array('id', 'username', 'race', 'elo', 'division')),
+								'Leader'=> array('fields' => array('id', 'username', 'race', 'elo', 'division')),
+								
+								)
+							));
+		$in_team = 0;
+		
+		foreach ($members as $members){
+			foreach ($members['User'] as $user){
+				if($user['id']==$this->Auth->user('id')){
+					$in_team = 1;
+				}
+			}
 		}
-		$this->redirect(array('action'=>'index'));
+		echo($in_team);					
+		if($in_team==1){
+			$this->Session->setFlash('Already part of the team!');
+			$this->redirect(array('controller'=>'teams','action'=>'view', $invitation['Invitation']['team_id']));
+		}
+		if($in_team==0){
+			$this->Invitation->Team->bindModel(array('hasMany'=>array('UsersTeams')));
+			$this->Invitation->Team->UsersTeams->create();
+			$this->data['UsersTeams']['team_id'] =  $invitation['Invitation']['team_id'];
+			$this->data['UsersTeams']['user_id'] = $invitation['Invitation']['user_id'];
+			$this->Invitation->delete();
+			if($this->Invitation->Team->UsersTeams->save ($this->data))
+			{
+				$this->Session->setFlash('Invitation accepted!');
+				//$this->redirect(array('controller'=>'teams','action'=>'view', $this->data['UsersTeams']['team_id']));
+			}
+			//$this->redirect(array('action'=>'index'));
+		}
+		
 		
 	}
 	
 	function decline ($id = null)
 	{
 		$this->Invitation->delete();
+	}
+	
+	function open_invitations()
+	{
+		
+		$current_user = $this->Session->read('Auth.User.id');
+		$open_invitations= $this->Invitation->find('count', array('conditions' => array('Invitation.user_id'=>$current_user)));
+		$this->set('open_invitations',$open_invitations);
+		
+		if (isset($this->params['requested'])) 
+		{     
+			return $open_invitations;        
+		}
 	}
 }
 ?>
